+++ fractal_server/app/schemas/v2/workflow.py
@@ -20,6 +20,7 @@ class WorkflowCreate(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     name: NonEmptyStr
+    description: NonEmptyStr | None = None
 
 
 class WorkflowRead(BaseModel):
@@ -45,7 +46,7 @@ class WorkflowUpdate(BaseModel):
 
     name: NonEmptyStr = None
     reordered_workflowtask_ids: ListUniqueNonNegativeInt | None = None
-    description: str | None = None
+    description: NonEmptyStr | None = None
 
 
 class WorkflowImport(BaseModel):
@@ -58,6 +59,7 @@ class WorkflowImport(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
     name: NonEmptyStr
+    description: NonEmptyStr | None = None
     task_list: list[WorkflowTaskImport]
 
 
@@ -70,4 +72,5 @@ class WorkflowExport(BaseModel):
     """
 
     name: str
+    description: str | None
     task_list: list[WorkflowTaskExport]